Add shared useDocsConfig hook
