+++ app/projects/table.tsx
@@ -1,13 +1,11 @@
-import { Link, useMatches, useNavigate, useParams } from '@remix-run/react'
-import { useMemo } from 'react'
+import { Link } from '@remix-run/react'
 import reactLogo from '~/images/react-logo.svg'
 import solidLogo from '~/images/solid-logo.svg'
 import vueLogo from '~/images/vue-logo.svg'
 import svelteLogo from '~/images/svelte-logo.svg'
-import type { AvailableOptions } from '~/components/Select'
-import type { ConfigSchema, MenuItem } from '~/utils/config'
-import { generatePath } from '~/utils/utils'
 import { FaDiscord, FaGithub } from 'react-icons/fa'
+import { useDocsConfig } from '~/utils/config'
+import type { ConfigSchema, MenuItem } from '~/utils/config'
 
 export const repo = 'tanstack/table'
 
@@ -74,109 +72,10 @@ export const createLogo = (version?: string) => (
 )
 
 export const useTableDocsConfig = (config: ConfigSchema) => {
-  const matches = useMatches()
-  const match = matches[matches.length - 1]
-  const params = useParams()
-  const version = params.version!
-  const framework = localStorage.getItem('framework') || 'react'
-  const navigate = useNavigate()
-
-  const frameworkMenuItems =
-    config.frameworkMenus?.find((d) => d.framework === framework)?.menuItems ??
-    []
-
-  const frameworkConfig = useMemo(() => {
-    if (!config.frameworkMenus) {
-      return undefined
-    }
-
-    const availableFrameworks = config.frameworkMenus?.reduce(
-      (acc: AvailableOptions, menuEntry) => {
-        if (menuEntry.framework in frameworks) {
-          acc[menuEntry.framework] =
-            frameworks[menuEntry.framework as keyof typeof frameworks]
-        }
-        return acc
-      },
-      { react: frameworks['react'] }
-    )
-
-    return {
-      label: 'Framework',
-      selected: framework,
-      available: availableFrameworks,
-      onSelect: (option: { label: string; value: string }) => {
-        const url = generatePath(match.id, {
-          ...match.params,
-          framework: option.value,
-        })
-
-        localStorage.setItem('framework', option.value)
-
-        navigate(url)
-      },
-    }
-  }, [config.frameworkMenus, framework, match, navigate])
-
-  const versionConfig = useMemo(() => {
-    const available = availableVersions.reduce(
-      (acc: AvailableOptions, version) => {
-        acc[version] = {
-          label: version,
-          value: version,
-        }
-        return acc
-      },
-      {
-        latest: {
-          label: 'Latest',
-          value: 'latest',
-        },
-      }
-    )
-
-    return {
-      label: 'Version',
-      selected: version,
-      available,
-      onSelect: (option: { label: string; value: string }) => {
-        const url = generatePath(match.id, {
-          ...match.params,
-          version: option.value,
-        })
-        navigate(url)
-      },
-    }
-  }, [version, match, navigate])
-
-  const docSearch: NonNullable<ConfigSchema['docSearch']> =
-    config.docSearch || {
-      appId: '',
-      apiKey: '',
-      indexName: '',
-    }
-
-  return {
-    ...config,
-    docSearch,
-    menu: [
-      localMenu,
-      // Merge the two menus together based on their group labels
-      ...config.menu.map((d) => {
-        const match = frameworkMenuItems.find((d2) => d2.label === d.label)
-        return {
-          label: d.label,
-          children: [
-            ...d.children.map((d) => ({ ...d, badge: 'core' })),
-            ...(match?.children ?? []).map((d) => ({ ...d, badge: framework })),
-          ],
-        }
-      }),
-      ...frameworkMenuItems.filter(
-        (d) => !config.menu.find((dd) => dd.label === d.label)
-      ),
-    ].filter(Boolean),
-    frameworkConfig,
-    versionConfig,
-  }
+  return useDocsConfig({
+    config,
+    frameworks,
+    localMenu,
+    availableVersions,
+  })
 }
